+++ mktlib/data/__init__.py
@@ -9,6 +9,7 @@ except ModuleNotFoundError as _e:
 
 from mktlib.data._gbm import geometric_brownian_motion
 from mktlib.data._monte_carlo import monte_carlo
+from mktlib.data._ohlcv import ticks_to_ohlcv
 from mktlib.data._ornstein_uhlenbeck import ornstein_uhlenbeck
 from mktlib.data._random_walk import fractional_random_walk
 
@@ -17,4 +18,5 @@ __all__ = [
     "geometric_brownian_motion",
     "monte_carlo",
     "ornstein_uhlenbeck",
+    "ticks_to_ohlcv",
 ]